+++ test/clj/stripe/charge_test.clj
@@ -5,8 +5,6 @@
             [stripe.test :as t]
             [stripe.test-data :as td]))
 
-(use-fixtures :once t/env-token)
-
 (deftest charge-test
   "Test for charging a customer."
   (t/with-customer [created td/customer-data]